[MSIC] Move to-string to core.
